requirements.txt
# --- LIBRERÍAS LIGERAS ---
fastapi
granian
pydantic
supabase
requests
python-jose[cryptography]
python-dotenv
httpx
litellm
slowapi
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
logtail-python
Pillow
piexif
cryptography
langdetect
sentry_sdk
redis==4.6.0
numpy
orjson
rapidfuzz

# --- MOTOR AI (VERSIÓN CPU OPTIMIZADA - < 300MB) ---
# Esto reduce el tamaño de la imagen en un 80%
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# --- SOVEREIGN AI (LOCAL INFERENCE) ---
onnxruntime>=1.18.0
tokenizers>=0.19.0
numpy
